JS: Created OneOrTwoKeyInput class
OneOrTwoKeyInput uses a HashMap with 2 keys where the second key is optional

When storing a value, if all two keys are known, then store the value with all two keys
Otherwise, store with the first key plus null.

When retrieving a value, first find the value for all two keys, but if it is not found,
then return the value for the first key plus null.

Example:

FPSO/Manager FractionOfCargoForBerthing { Panamax 0.7 } "for all product types
or
FPSO/Manager FractionOfCargoForBerthing { Panamax Propane 0.6 } "for Propane only

Signed-off-by: Stephen Wong <stephen.wong@ausencosandwell.com>
Signed-off-by: Harvey Harrison <harvey.harrison@ausencosandwell.com>