JS, JS3D: Added Entity method updateForInput that updates the Entity for changes in the given input
This method is called immediately after parsing the data for the Input object

This change will allow objects to update immediately after the input has been read
from the input file or set in the EditBox.

For example, DisplayEntities will update their position immediately after the Position input has been read
from the input file or set in the EditBox.

Signed-off-by: Stephen Wong <stephen.wong@ausencosandwell.com>
Signed-off-by: Harvey Harrison <harvey.harrison@ausencosandwell.com>