JS3D: Moved code for Region.initializeLocalClock to constructor and render methods.
Previously, the localClock was created after user loaded a configuration file.
This caused a null pointer error when the inputs were embedded in the jar file and
the user did not load a configuration file (because localClock was still null).

Added bonus: The user can edit the ShowTime input to toggle a region's status bar.

Signed-off-by: Stephen Wong <stephen.wong@ausencosandwell.com>
Signed-off-by: Harvey Harrison <harvey.harrison@ausencosandwell.com>