JS: Fixed FileEntity toStart
inputStream.reset() goes to the last marked position of the file, which may not be the start.
This caused the metocean file to loop back to the last marked position, instead of the start of the file,
as there is one place in the code where we mark the position in the metocean file, then look forward, and then go back to the marked position.

Signed-off-by: Stephen Wong <stephen.wong@ausencosandwell.com>
Signed-off-by: Harvey Harrison <harvey.harrison@ausencosandwell.com>