JS3D: fix not saving some interactively changed input when saving to an input file
Parse the value to the input rather than updating just their value strings.

After the value changes interactively, it is put  in a string. then, the
program parses the string and put it back to the value. This seems
inefficient, but we don't want to change the Value input directly for
safety reasons.

Signed-off-by: Ali Marandi <ali.marandi@ausencosandwell.com>
Signed-off-by: Harvey Harrison <harvey.harrison@ausencosandwell.com>