JS3D: Added TextLabel keywords DropShadow and DropShadowColour
The DropShadow keyword allows the user to easily add a drop shadow to a text label without having to define a second text label with an offset.

DropShadow - if the value is true, then a drop shadow appears for the text label.  Default is false.

DropShadowColour - the colour of the drop shadow.  Default is black.

Example:

SimulationTitle DropShadow { TRUE }  DropShadowColour { red }

Signed-off-by: Stephen Wong <stephen.wong@ausencosandwell.com>
Signed-off-by: Harvey Harrison <harvey.harrison@ausencosandwell.com>