JS3D: add the ability to edit ListInput for input editor
There were two ways of doing this:

1) Customize dropdown list we already had to work for ListInput as well
2) Create a new editor for ListInput

For the first option we just need to render the items in our combobox to
JCheckBox and have a mouse adapter to listen to mouse click. This was not a
good solution because the user can select only one item at the time.

The second solution (which is taken) is a JPanal with a text and a button
inside just like ColorEditor. When the user clicks on the button, a model
dialog box will be opened with a list of all the possible options. If there
was already a value in the input, those items from the input value are get
check marked inside the list. Then, the user is able to select/unselect
items. If the user presses the "OK" button at the bottom of the dialog box,
the selection is applied to the input. Otherwise, it won’t have any effect.
The user still have the option to enter the list manually.

Signed-off-by: Ali Marandi <ali.marandi@ausencosandwell.com>
Signed-off-by: Harvey Harrison <harvey.harrison@ausencosandwell.com>