JS, JS3D: add pauseAt TextField to mainToolbar to run the model to the time
The user can enter the time in the form of hours or date yyyy-mm-ddd. The
default is infinity. If they press the run button, the program will pause
at that time.

When the pauseTime looses the focus and it is empty, it shows a infinity
sign which means there is no pause time.

remove the ugly hack because it is already done in startSimulation method.

Signed-off-by: Ali Marandi <ali.marandi@ausencosandwell.com>
Signed-off-by: Harvey Harrison <harvey.harrison@ausencosandwell.com>