JS3D: define rotating cursor for interactively rotating an Entity
getBestCursorSize returns the best size for the image.

Note: if an image is used whose dimensions don't match a supported size (as
returned by getBestCursorSize method), the Toolkit implementation will
attempt to resize the image to a supported size which does not look the
best.

getBestCursorSize returns a dimension of 0, 0 for systems which don't
support custom cursors.

Signed-off-by: Ali Marandi <ali.marandi@ausencosandwell.com>
Signed-off-by: Harvey Harrison <harvey.harrison@ausencosandwell.com>