TLS, JS3D: use state record to obtain total hours spent in a given state
for printing to the UTL report

totalHoursPerState did not contain the hours up to the parking block. It
also did not contain the current cycle hours up to the initialization.
However, totalHours from state record contains those hours. This might
cause some change to Signal Block Group section in UTL report where this
method is used.

Signed-off-by: Ali Marandi <ali.marandi@ausenco.com>
Signed-off-by: Harvey Harrison <harvey.harrison@ausenco.com>