JS: move testing of the die flag into threadWait
This allows us to kill Processes that are waiting in a chain, as created with interrupt
or startProcess.  Now when an error is propagated, we properly tear down these live
threadStacks that used to be left in place, preventing restart from being workable.

We now clear processRunning in the error handling, so that restart will be allowed
to run new events.

Signed-off-by: Harvey Harrison <harvey.harrison@ausenco.com>