JS: restore the ordering of inputs in the EditBox
We relied on the ordering of the editable keywords to be in the order they were
added, using an iterator over a hashmap returns essentially a random order.  Add
a list specifically to hold the input objects in the needed order.

This has the benefit of providing a unique list of all supported inputs as the
synonym mappings will not be added here.

Signed-off-by: Harvey Harrison <harvey.harrison@ausenco.com>