JS: remove the Choice keyword for Assemble
The ability to select the right queue for an incoming entity using the
Choice keyword will not work once the output obj becomes the entity being
processed instead of the entity last received. In any case, it is more
logical for incoming entities to be send directly to the appropriate
queue.

Signed-off-by: Harry King <harry.king@ausenco.com>
Signed-off-by: Harvey Harrison <harvey.harrison@ausenco.com>