JS: pull the target ticks into the start method
This avoids the case where we failed validation, but end up with an empty event
queue in the EventManager, which then makes the model somewhat non-functional
as we just try to resume the model each time the start button is pressed instead
of reinitializing the whole thing.

As such, when validation fails, mark the model as configured, but not started.

Signed-off-by: Harvey Harrison <harvey.harrison@ausenco.com>