JS: replace the SimulationTimeScale keyword with TickLength
Specify the length of time represented by one 'tick' as the primary input. Derive
the reciprocal value from that, knowing that some precision will be lost in the
division, but this should only be used when we intend to round off to a whole tick
anyway, which avoids lost ticks.

Also, leave the default as microseconds and do not change this in the default file.

Signed-off-by: Harvey Harrison <harvey.harrison@ausenco.com>