JS: reset input state to valid for DropDownMenuEditor
Required because stopCellEditing is called when the cell is selected by the
user as well as when editing is finished. If an error message is generated
and the user selects the "edit" option, then the error message is simply
regenerated, preventing the user from fixing the input. Only the
DropDownMenuEditor has this problem - the other editors call stopCellEditing
only when editing is finished.

Signed-off-by: Harry King <d.harry.king@gmail.com>