JS: correct the code that re-opens the log file for the simulation run
The log file is not retained at the end of the run unless there is a
non-zero error or warning count. None of the logged messages during the run
increment these counts, so the file is always closed. The code has been
corrected so that the log file could be used in the future.

Signed-off-by: Harry King <d.harry.king@gmail.com>