JS: save a copy of the ExpressionThreshold state when it is calculated on-demand
Fixes a bug that occurs when the ExpressionThreshold changes state twice at
the same simulation time. Only the second state change is captured by
Threshold's "open" property. This is okay except in the situation where
both the OpenCondition and CloseCondition inputs are false, in which case
the state of the ExpressionThreshold is unchanged. Prior to this commit,
the unchanged state was determine by super.isOpen(), which did not provide
a reliable value. The unchanged state is now maintained directly from the
previous value that was calculated.

Signed-off-by: Harry King <d.harry.king@gmail.com>