JS: add # as an alternate line-commenting character
It is safe to remoe the test for quoted tokens in echoInputRecord as all
existing callers strip comments before calling it, so this should never be
getting hit. It is also unsafe to test for quoting this way as the
precedence has changed, and charaters such as " can exist in legitimate
tokens.

Signed-off-by: Harvey Harrison <harvey.harrison@gmail.com>