JS: add work around to scheduleExternalEvent scheduling next eventi
These events can be injected from external code during a real-time wait,
which possibly moves the next event execution time backwards which was
never anticipated previously. Add aworkaround to move the nextTick value
back to be consistent, we do not also need to validate currentTick as
calculateEventTime() has already done so.

Possibly useful to remove the nextTick concept in future and always just
test the leftmost node in the RB-tree, but this will do for now.

Signed-off-by: Harvey Harrison <harvey.harrison@gmail.com>