JS: remove validation for min and max value from SampleInput and SampleListInput
Cannot define reliable minimum and maximum values for a SampleExpression.
If getMinValue and getMaxValue return the present value for the expression
instead of plus/minus infinity, then there are cases when the expression
cannot be evaluated, for example when it includes 'this.obj' which is null.
It is better to remove the validation instead of trying to patch it to work
some of the time.

Signed-off-by: Harry King <d.harry.king@gmail.com>