JS: add flag 'REGISTERED' to Entity
This change distinguishes between a 'generated' entity and a 'registered'
entity. A generated entity is one that is created during the simulation
run and is destroyed at the end of the run. A registered entity is one that
is included in the namedEntities hashmap, which allows other entities to
reference it by name. Only registered entities are able to appear in the
inputs of other entities. Up to now, all generated entities were assumed to
be unregistered.

Signed-off-by: Harry King <d.harry.king@gmail.com>