JS: fix null pointer crashes in updateForInput for GUI keywords
Fixes: de79a479a72e64754c2c0c8bec42373f6df99b5b

This commit does not ensure there is a default value in the inputs in
the headless case, where Simulation.setWindowDefaults() does not get
called. If any of these keywords are used, they cause null pointer
errors in the updateForInput callback.

The crash comes from within the constructor of the window created by a
call to getInstance() which creates one if necessary, in the headless
case, none have been created before we read inputs, and we crash trying
to set the initial location or size of the window reading the null
default value of the Simulation's inputs.

Signed-off-by: Harvey Harrison <harvey.harrison@gmail.com>