JS: discourage toggle buttons from taking the focus
The background colour used to show focus makes a toggle button look like
it is activated. It is also turned off for the undo/redo buttons because
this also makes it less likely that the toggle buttons to the right will
take the focus if the undo/redo buttons become disabled.

Signed-off-by: Harry King <d.harry.king@gmail.com>