JS: prevent an entity name from including reserved characters
The list of invalid characters has been increased to include single and
double quotes, square brackets, and the hash character. An entity name that
includes any of these characters, except for the double quotes, causes a
saved input file to be parsed incorrectly when it is reloaded. It appears
that an entity name could include a double quote, but this character has
also been excluded to avoid any future problems.

Signed-off-by: Harry King <d.harry.king@gmail.com>