JS: modify getLabel to find an entity's EntityLabel based on its name
This method is less robust than using the TargetEntity input for the
EntityLabel, however, it allows the input file to be loaded correctly.
The Simulation keyword ShowLabels is processed before the TargetEntity for
the EntityLabels, which causes a second EntityLabel to be created when
getLabels relies on the TargetEntity value. Another way to solve the
problem would be to make TargetEntity an 'early' keyword, but this would
not fix the existing input files.

Signed-off-by: Harry King <d.harry.king@gmail.com>