JS: modify updateGraphics for Threshold
A Threshold that is the component of a sub-model needs to be shown/hidden
based on the 'Show SubModels' button which uses setShow. This mechanism is
incompatible with the ShowWhenOpen and ShowWhenClosed keywords. The problem
is resolved for now by ignoring these inputs if they have their default
values (TRUE).

Signed-off-by: Harry King <d.harry.king@gmail.com>