JS: use CountDownLatch in TestSimulation unit tests
Fixes a bug in waitForPause that occurs because the present system cannot
distinguish between the following two cases when waitForPause is called
and isRunning is found to be false: either (1) event processing has not
started yet, or (2) the simulation has finished already.

The CountDownLatch solves this problem because its count is set to one on
creation and can be decremented before waitForPause is called.

Signed-off-by: Harry King <d.harry.king@gmail.com>