JS: making locking consistent within inner EventManager methods
All the internal scheduling methods that effect event state are only
valid to be called on the current eventmanager from a running process,
this is enforced by putting all callers through the static methods that
look up the current running Process and its EventManager.

A running Process is guaranteed to already hold the lockObject as it
will have passed through the execute() method and will still hold the
lock. The scheduleTicks method had already had this done, normalize the
start, interrupt, kill, scheduleUntil, wait and waitUntil methods to match
the locking in scheduleWait.

Signed-off-by: Harvey Harrison <harvey.harrison@gmail.com>