JS: update the value in the selected cell for the Input Editor
Fixes a bug that prevented the value in the selected cell from changing
when the Undo button was clicked. This bug affected the earlier version of
the Input Editor, but was more noticeable after the change to the new
spreadsheet style of inputs where a cell can be selected without being in
edit mode.

Signed-off-by: Harry King <d.harry.king@gmail.com>