JS: retain the retryString in CellEditor
Fixes a bug introduced by 7288340bc034c2624a4094124ca62a768e383280
(JS: update the value in the selected cell for the Input Editor) that
prevents the 'Edit' option from working correctly when an input error is
made. Note that it is not necessary to reset the retryString property to
null because the CellEditor is discarded after editing is stopped.

Signed-off-by: Harry King <d.harry.king@gmail.com>