JS: Add DowntimeEntity keyword 'CompletionTimeLimit' and Output 'LateEvents'
The total time from the scheduled start time that the downtime event should be completed within.
For example, if the scheduled start time is 100 h and the completion time limit is 48 h, the event
will be recorded as late in the 'LateEvents' output if it is not completed by 148h.

Signed-off-by: Geoffrey Savinkoff <geoffrey.savinkoff@ausenco.com>