JS: increase tolerance for probability distribution unit tests
Fixes unit test failures caused by commit
febea692e053ad4a9218d1af891c89997b4ec24d
(JS: set the default input for the 'GlobalSubstreamSeed' keyword to the
present replication number)

This commit changed the default GlobalSubstreamSeed value from 0 to 1,
which caused the sampled values to change for each distribution. The former
tolerances were fine tuned for the original GlobalSubstreamSeed value.

A tolerance of 0.01 was chosen to be large enough for most distributions.
Note that by making 1,000,000 samples, the confidence interval for the
mean value is reduced by a factor of 0.001.

Signed-off-by: Harry King <d.harry.king@gmail.com>