JS: improve synchronization of Scenario methods
Prevents a deadlock occuring when Scenario.runEnded is called while
RunManager.start is still executing and is trying to start a new run for
same scenario as the run that completed. The change to runEnded prevents
a thread from taking the lock for RunManager while it still has the lock
for its scenario.

Signed-off-by: Harry King <d.harry.king@gmail.com>

# Conflicts:
#	src/main/java/com/jaamsim/basicsim/Scenario.java

# Conflicts:
#	src/main/java/com/jaamsim/basicsim/Scenario.java