JS: ensure that ObjectType inputs are not saved to the configuration file
This should never happen but there have been cases where the software has
crashed during configuration file saving because an entity's object type
was null. Prior to this commit, this type of error could happen if an
ObjectType entity was somehow marked as edited. Since getObjectType returns
null for an instance of ObjectType, a null pointer error would occur when
the list of entities was sorted by uiEntitySortOrder.

Signed-off-by: Harry King <d.harry.king@gmail.com>