JS: fix deadlock in validation error dialog
Fixes a bug introduced by a6b7b5950ca5b7c01d0f566ed351e026375ba22f
(JS: update the RunProgressBox through GUIFrame.updateUI) that caused the
Swing thread to deadlock when an error occured during validation in a model
that had not been saved already. When this occurred, only part of the error
message popup appeared and it could not be closed in the normal fashion.

Signed-off-by: Harry King <d.harry.king@gmail.com>