JS: use the runningProc member to control event loop
If we are still the runningProc after executing an event, continue
executing events. If not, we are a thread that completed an event after
being woken up, a startProcess/interrupt thread, or a killed thread and
we shoudl exit back to the pool. Simplifies the odd calling convention
that essentially returned whether or not there was a next thread that
would continue event execution for us. Also makes sure the runningProc
is up to date on all Thread transitions.

Signed-off-by: Harvey Harrison <harvey.harrison@gmail.com>