JS: factor out common code to pushThread method
Event start, startProcess and interrupt all use the same pieces factor this out
based on the interrupt path which has to handle both allocating a process or
using a captured process.

This allows remving the threadWait method as the waiting logic for
passing thread control can be inlined in the new pushthread helper.

Signed-off-by: Harvey Harrison <harvey.harrison@gmail.com>