JS: avoid null pointer error while traversing entity list
Have to be careful traversing the list in case kill removes the next
entity in the traversal, make sure to skip enough deleted nodes to get
back to the live nodes.

Both the reset path clearing generated entities and the overall clear
method are adjusted to follow the same pattern.

Signed-off-by: Harvey Harrison <harvey.harrison@gmail.com>