JS: add a return code to executeTarget to avoid testing runningProc
We know within execute target whether the current thread should exit or
not, add that as a return code and avoid testing the runningProc in
execute(), this can be null now when an error occurs during the model
run and we end up with a null pointer exception. This is also slightly
more efficient as we avoid a test for runningProc.get().proc in the main
event loop.

Signed-off-by: Harvey Harrison <harvey.harrison@gmail.com>